+++ src/services/index.ts
@@ -1,29 +1,37 @@
-import axios from './config'
+import axios from "./config";
 
-// export const SERVER_URL = 'http://localhost:5000'
-export const SERVER_URL = (import.meta.env.MODE === 'development') ? '/api' : 'http://127.0.0.1:5001/api'
-export const ASSET_URL = 'http://127.0.0.1:5001/assets'
-export const API_URL = `http://127.0.0.1:5001/api`;
+
+
+export const SERVER_URL =
+  import.meta.env.MODE === "development" ? "/api" : "http://127.0.0.1:3001/api";
+export const ASSET_URL = "http://127.0.0.1:3001/assets";
+export const API_URL = `http://127.0.0.1:3001/api`;
+
+
+const url = new URL(window.location.href);
+const t = url.searchParams.get("t") ? `?t=${url.searchParams.get("t")}` : ``;
 
 
 export default {
   getMockData(filename: string): Promise<any> {
-    return axios.get(`./mocks/${filename}.json`)
+    return axios.get(`./mocks/${filename}.json`);
   },
 
   getFileData(filename: string): Promise<any> {
-    return axios.get(`${API_URL}/documents/${filename}`);
+    const e = `${API_URL}/documents/${filename}${t}`;
+    return axios.get(e);
   },
-  
+
+
   AIPPT_Outline(
     content: string,
     language: string,
-    model: string,
+    model: string
   ): Promise<any> {
     return fetch(`${SERVER_URL}/tools/aippt_outline`, {
-      method: 'POST',
+      method: "POST",
       headers: {
-        'Content-Type': 'application/json',
+        "Content-Type": "application/json",
       },
       body: JSON.stringify({
         content,
@@ -31,18 +39,14 @@ export default {
         model,
         stream: true,
       }),
-    })
+    });
   },
 
-  AIPPT(
-    content: string,
-    language: string,
-    model: string,
-  ): Promise<any> {
+  AIPPT(content: string, language: string, model: string): Promise<any> {
     return fetch(`${SERVER_URL}/tools/aippt`, {
-      method: 'POST',
+      method: "POST",
       headers: {
-        'Content-Type': 'application/json',
+        "Content-Type": "application/json",
       },
       body: JSON.stringify({
         content,
@@ -50,6 +54,6 @@ export default {
         model,
         stream: true,
       }),
-    })
+    });
   },
-}
+};