+++ vite.config.ts
@@ -5,6 +5,9 @@ import vue from '@vitejs/plugin-vue'
 
 // https://vitejs.dev/config/
 export default defineConfig({
+  build: {
+    chunkSizeWarningLimit: 10000,
+  },
   base: '',
   plugins: [
     vue(),